Cards refactor
The logic behind the contents of the card has been moved to a Card
component which renders a proper body based on the past card.
